fix: rename folder name + uninstall @material-ui/core + use new way instead of makeStyles
