+++ client/package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
-    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.10.6",
     "@mui/material": "^5.10.8",
     "@testing-library/jest-dom": "^5.16.5",
