+++ client/src/index.css
@@ -3,7 +3,7 @@
 
 :root {
   --main-color: #0A20E6;
-  --text-color: #424B5A;
+  --text-color: #2f3643;
   --second-text-color: #A6AFBA;
   --container-color: #FFF;
   --background-color: #F6F7F9;
